refact: rename `mail_outbox` resource to `email_outbox` across provider and docs
