+++ docs/resources/email_outbox.md
@@ -1,0 +1,44 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "mittwald_email_outbox Resource - terraform-provider-mittwald"
+subcategory: ""
+description: |-
+  This resource models a mail outbox.
+---
+
+# mittwald_email_outbox (Resource)
+
+This resource models a mail outbox.
+
+## Example Usage
+
+```terraform
+resource "random_password" "mailbox_password" {
+  length           = 16
+  special          = true
+  override_special = "!#$%&*()-_=+[]{}<>:?"
+}
+
+resource "mittwald_email_outbox" "example" {
+  project_id  = "p-12345"
+  description = "Example mail outbox"
+  password    = resource.random_password.mailbox_password.result
+}
+
+output "email_outbox_id" {
+  value = mittwald_email_outbox.example.id
+}
+```
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Required
+
+- `description` (String) The description of the email outbox.
+- `password` (String, Sensitive) The password for the email outbox. For security, it is recommended to use the 'random_password' Terraform resource to dynamically generate a secure password instead of hardcoding values.
+- `project_id` (String) The ID of the project the email outbox belongs to
+
+### Read-Only
+
+- `id` (String) The generated email outbox ID
