--- docs/resources/mail_outbox.md
@@ -1,44 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "mittwald_mail_outbox Resource - terraform-provider-mittwald"
-subcategory: ""
-description: |-
-  This resource models a mail outbox.
----
-
-# mittwald_mail_outbox (Resource)
-
-This resource models a mail outbox.
-
-## Example Usage
-
-```terraform
-resource "random_password" "mailbox_password" {
-  length           = 16
-  special          = true
-  override_special = "!#$%&*()-_=+[]{}<>:?"
-}
-
-resource "mittwald_mail_outbox" "example" {
-  project_id  = "p-12345"
-  description = "Example mail outbox"
-  password    = resource.random_password.mailbox_password.result
-}
-
-output "mail_outbox_id" {
-  value = mittwald_mail_outbox.example.id
-}
-```
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `description` (String) The description of the mail outbox.
-- `password` (String, Sensitive) The password for the mail outbox. For security, it is recommended to use the 'random_password' Terraform resource to dynamically generate a secure password instead of hardcoding values.
-- `project_id` (String) The ID of the project the mail_outbox belongs to
-
-### Read-Only
-
-- `id` (String) The generated mail_outbox ID
