+++ examples/resources/mittwald_email_outbox/resource.tf
@@ -4,12 +4,12 @@ resource "random_password" "mailbox_password" {
   override_special = "!#$%&*()-_=+[]{}<>:?"
 }
 
-resource "mittwald_mail_outbox" "example" {
+resource "mittwald_email_outbox" "example" {
   project_id  = "p-12345"
   description = "Example mail outbox"
   password    = resource.random_password.mailbox_password.result
 }
 
-output "mail_outbox_id" {
-  value = mittwald_mail_outbox.example.id
+output "email_outbox_id" {
+  value = mittwald_email_outbox.example.id
 }
