+++ internal/provider/resource/emailoutboxresource/model_api.go
@@ -1,4 +1,4 @@
-package mailoutboxresource
+package emailoutboxresource
 
 import (
 	"context"
